Configure project for Search in Rotated Sorted Array
-	Created new Solution class
-	Create new test call with failing test for a non rotated array
